Module6-task3 промежуточная реализация 1.1
